refactor: WinningNumber 제거, Lotto 클래스로 대체.
 - 가지고 있는 필드가 동일해 대체.
