+++ src/main/java/lotto/LottoApplication.java
@@ -8,7 +8,6 @@ import lotto.domain.LottoNumber;
 import lotto.domain.LottoRanks;
 import lotto.domain.Lottos;
 import lotto.domain.Statistics;
-import lotto.domain.WinningNumbers;
 import lotto.domain.dto.ResultDto;
 import lotto.view.InputView;
 import lotto.view.PrintView;
@@ -23,7 +22,7 @@ public class LottoApplication {
         PrintView.printLottoNumber(lottos);
 
         final CorrectNumbers correctNumbers = new CorrectNumbers(
-                new WinningNumbers(InputView.inputWinningNumbers()),
+                InputView.inputWinningNumbers(),
                 new LottoNumber(InputView.inputBonusNumber()));
 
         final Judge judge = new Judge(correctNumbers);
